requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
python-telegram-bot==20.4
scikit-learn==1.3.0
schedule==1.2.0
beautifulsoup4==4.12.2
selenium==4.11.2
aiohttp==3.8.5
asyncio==3.4.3
sqlalchemy==2.0.20
python-dotenv==1.0.0
joblib==1.3.2
pytest==7.4.0
pytest-asyncio==0.21.0
gunicorn==21.2.0